third_party/WebKit/LayoutTests/animations/svg-attribute-composition/svg-transform-composition-distinct-expected.txt

This is a testharness.js-based test.
PASS This test uses interpolation-test.js. 
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewX(10)] from add [translate(10 20)] to add [rotate(30 40 50)] at (-0.4) is [skewX(10) translate(10 20)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 10 , 20 " but got "2 , 0 , 1 , 0 , 0 , 1 , 10 , 20 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewX(10)] from add [translate(10 20)] to add [rotate(30 40 50)] at (0) is [skewX(10) translate(10 20)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 10 , 20 " but got "2 , 0 , 1 , 0 , 0 , 1 , 10 , 20 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewX(10)] from add [translate(10 20)] to add [rotate(30 40 50)] at (0.2) is [skewX(10) translate(10 20)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 2 , 0 , 1 , 0 , 0 , 1 , 10 , 20 " but got "2 , 0 , 1 , 0 , 0 , 1 , 10 , 20 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewX(10)] from add [translate(10 20)] to add [rotate(30 40 50)] at (0.6) is [skewX(10) rotate(30 40 50)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 " but got "4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewX(10)] from add [translate(10 20)] to add [rotate(30 40 50)] at (1) is [skewX(10) rotate(30 40 50)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 " but got "4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewX(10)] from add [translate(10 20)] to add [rotate(30 40 50)] at (1.4) is [skewX(10) rotate(30 40 50)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 " but got "4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from add [translate(5 15)] to add [rotate(30 40 50)] at (-0.4) is [translate(5 5) translate(5 15)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 " but got "2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from add [translate(5 15)] to add [rotate(30 40 50)] at (0) is [translate(5 5) translate(5 15)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 " but got "2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from add [translate(5 15)] to add [rotate(30 40 50)] at (0.2) is [translate(5 5) translate(5 15)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 , 2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 " but got "2 , 0 , 1 , 0 , 0 , 1 , 5 , 15 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from add [translate(5 15)] to add [rotate(30 40 50)] at (0.6) is [translate(5 5) rotate(30 40 50)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 " but got "4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from add [translate(5 15)] to add [rotate(30 40 50)] at (1) is [translate(5 5) rotate(30 40 50)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 " but got "4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from add [translate(5 15)] to add [rotate(30 40 50)] at (1.4) is [translate(5 5) rotate(30 40 50)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 , 4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 " but got "4 , 30 , 0.87 , 0.5 , - 0.5 , 0.87 , 30.36 , - 13.3 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewY(5)] from add [skewX(10)] to add [skewY(15)] at (-0.4) is [skewY(5) skewX(10)] assert_equals: expected "6 , 5 , 1 , 0.09 , 0 , 1 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewY(5)] from add [skewX(10)] to add [skewY(15)] at (0) is [skewY(5) skewX(10)] assert_equals: expected "6 , 5 , 1 , 0.09 , 0 , 1 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewY(5)] from add [skewX(10)] to add [skewY(15)] at (0.2) is [skewY(5) skewX(10)] assert_equals: expected "6 , 5 , 1 , 0.09 , 0 , 1 , 0 , 0 , 5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 " but got "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewY(5)] from add [skewX(10)] to add [skewY(15)] at (0.6) is [skewY(5) skewY(15)] assert_equals: expected "6 , 5 , 1 , 0.09 , 0 , 1 , 0 , 0 , 6 , 15 , 1 , 0.27 , 0 , 1 , 0 , 0 " but got "6 , 15 , 1 , 0.27 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewY(5)] from add [skewX(10)] to add [skewY(15)] at (1) is [skewY(5) skewY(15)] assert_equals: expected "6 , 5 , 1 , 0.09 , 0 , 1 , 0 , 0 , 6 , 15 , 1 , 0.27 , 0 , 1 , 0 , 0 " but got "6 , 15 , 1 , 0.27 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [skewY(5)] from add [skewX(10)] to add [skewY(15)] at (1.4) is [skewY(5) skewY(15)] assert_equals: expected "6 , 5 , 1 , 0.09 , 0 , 1 , 0 , 0 , 6 , 15 , 1 , 0.27 , 0 , 1 , 0 , 0 " but got "6 , 15 , 1 , 0.27 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from neutral to add [skewY(20)] at (-0.4) is [translate(5 5)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from neutral to add [skewY(20)] at (0) is [translate(5 5)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from neutral to add [skewY(20)] at (0.2) is [translate(5 5)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from neutral to add [skewY(20)] at (0.6) is [translate(5 5) skewY(20)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 , 6 , 20 , 1 , 0.36 , 0 , 1 , 0 , 0 " but got "6 , 20 , 1 , 0.36 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from neutral to add [skewY(20)] at (1) is [translate(5 5) skewY(20)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 , 6 , 20 , 1 , 0.36 , 0 , 1 , 0 , 0 " but got "6 , 20 , 1 , 0.36 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <transform> with underlying [translate(5 5)] from neutral to add [skewY(20)] at (1.4) is [translate(5 5) skewY(20)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 5 , 5 , 6 , 20 , 1 , 0.36 , 0 , 1 , 0 , 0 " but got "6 , 20 , 1 , 0.36 , 0 , 1 , 0 , 0 "
Harness: the test ran to completion.
